+++ apps/meteor/lib/callbacks/callbacksBase.ts
@@ -170,7 +170,7 @@ export class Callbacks<
 		this.setCallbacks(hook, hooks);
 	}
 
-	run<Hook extends keyof TEventLikeCallbackSignatures>(hook: Hook, ...args: Parameters<TEventLikeCallbackSignatures[Hook]>): void;
+	run<Hook extends keyof TEventLikeCallbackSignatures>(hook: Hook, ...args: Parameters<TEventLikeCallbackSignatures[Hook]>): Promise<void>;
 
 	run<Hook extends keyof TChainedCallbackSignatures>(
 		hook: Hook,