+++ apps/meteor/server/services/omnichannel/queue.ts
@@ -1,4 +1,5 @@
 import type { InquiryWithAgentInfo, IOmnichannelQueue } from '@rocket.chat/core-typings';
+import { License } from '@rocket.chat/license';
 import { LivechatInquiry } from '@rocket.chat/models';
 
 import { dispatchAgentDelegated } from '../../../app/livechat/server/lib/Helper';
@@ -19,25 +20,28 @@ export class OmnichannelQueue implements IOmnichannelQueue {
 		return timeout < 1 ? DEFAULT_RACE_TIMEOUT : timeout * 1000;
 	}
 
+	public isRunning() {
+		return this.running;
+	}
+
 	async start() {
-		queueLogger.debug('Starting queue');
 		if (this.running) {
-			queueLogger.debug('Queue already running');
 			return;
 		}
 
 		const activeQueues = await this.getActiveQueues();
 		queueLogger.debug(`Active queues: ${activeQueues.length}`);
-
 		this.running = true;
+
+		queueLogger.info('Service started');
 		return this.execute();
 	}
 
 	async stop() {
-		queueLogger.debug('Stopping queue');
 		await LivechatInquiry.unlockAll();
 
 		this.running = false;
+		queueLogger.info('Service stopped');
 	}
 
 	private async getActiveQueues() {
@@ -62,7 +66,7 @@ export class OmnichannelQueue implements IOmnichannelQueue {
 
 		const queue = await this.nextQueue();
 		const queueDelayTimeout = this.delay();
-		queueLogger.debug(`Executing queue ${queue || 'Public'} with timeout of ${queueDelayTimeout}`);
+		queueLogger.info(`Executing queue ${queue || 'Public'} with timeout of ${queueDelayTimeout}`);
 
 		setTimeout(this.checkQueue.bind(this, queue), queueDelayTimeout);
 	}
@@ -95,9 +99,13 @@ export class OmnichannelQueue implements IOmnichannelQueue {
 		}
 	}
 
-	shouldStart() {
+	async shouldStart() {
 		if (!settings.get('Livechat_enabled')) {
 			void this.stop();
+			return;
+		}
+
+		if (await License.shouldPreventAction('monthlyActiveContacts')) {
 			return;
 		}