+++ packages/ui-video-conf/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rocket.chat/ui-video-conf
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies [074db3b419]
+- Updated dependencies [b8f3d5014f]
+  - @rocket.chat/ui-contexts@2.0.0
+
+## 2.0.0-rc.5
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@2.0.0-rc.5
+
 ## 2.0.0-rc.4
 
 ### Patch Changes
@@ -31,6 +45,12 @@
 - Updated dependencies [074db3b419]
 - Updated dependencies [b8f3d5014f]
   - @rocket.chat/ui-contexts@2.0.0-rc.0
+
+## 1.0.8
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@1.0.8
 
 ## 1.0.7